chore: refactor dsat module to be independent of deepspeed imports [MLG-499]
