+++ ds_autotuning_prototype/core_api/ffn_example_ds_launcher/script.py
@@ -7,7 +7,7 @@ import torch
 import torch.nn as nn
 import torch.nn.functional as F
 from attrdict import AttrDict
-from determined.pytorch.deepspeed import dsat, get_ds_config_from_hparams
+from determined.pytorch import dsat
 from torch.utils.data import Dataset
 
 
@@ -18,7 +18,7 @@ class RandDataset(Dataset):
         self.data = torch.randn(self.num_actual_datapoints, self.dim)
 
     def __len__(self) -> int:
-        return 10 ** 6
+        return 10**6
 
     def __getitem__(self, idx: int) -> torch.Tensor:
         data = self.data[idx % self.num_actual_datapoints]
@@ -48,7 +48,7 @@ def main(
     hparams = AttrDict(hparams)
     if is_chief:
         logging.info(f"HPs seen by trial: {hparams}")
-    ds_config = get_ds_config_from_hparams(hparams)
+    ds_config = dsat.get_ds_config_from_hparams(hparams)
     dataset = RandDataset(hparams.dim)
     model = MinimalModel(hparams.dim, hparams.layers)
