+++ ds_autotuning_prototype/core_api/ffn_example_torch_launcher/script.py
@@ -7,7 +7,7 @@ import torch
 import torch.nn as nn
 import torch.nn.functional as F
 from attrdict import AttrDict
-from determined.pytorch.deepspeed import dsat, get_ds_config_from_hparams
+from determined.pytorch import dsat
 from torch.utils.data import Dataset
 
 
@@ -18,7 +18,7 @@ class RandDataset(Dataset):
         self.data = torch.randn(self.num_actual_datapoints, self.dim)
 
     def __len__(self) -> int:
-        return 10 ** 6
+        return 10**6
 
     def __getitem__(self, idx: int) -> torch.Tensor:
         data = self.data[idx % self.num_actual_datapoints]
@@ -49,7 +49,7 @@ def main(
         logging.info(f"HPs seen by trial: {hparams}")
     # Hack for clashing 'type' key. Need to change config parsing behavior so that
     # user scripts don't need to inject helper functions like this.
-    ds_config = get_ds_config_from_hparams(hparams)
+    ds_config = dsat.get_ds_config_from_hparams(hparams)
     dataset = RandDataset(hparams.dim)
     model = MinimalModel(hparams.dim, hparams.layers)
