+++ ds_autotuning_prototype/core_api/no_op_test/deepspeed.yaml
@@ -2,10 +2,6 @@ name: no_op_test_dsat
 workspace: DS_AT_PROTOTYPING
 project: no_op_test
 max_restarts: 0
-environment:
-  image:
-    gpu: determinedai/environments:cuda-11.3-pytorch-1.10-deepspeed-0.8.3-gpu-0.21.2
-    cpu: determinedai/environments:cuda-11.3-pytorch-1.10-deepspeed-0.8.3-gpu-0.21.2
 resources:
   slots_per_trial: 2
   # max_slots: 8
