+++ ds_autotuning_prototype/core_api/torchvison_models/deepspeed.yaml
@@ -5,7 +5,6 @@ max_restarts: 0
 environment:
   image:
     gpu: determinedai/environments:cuda-11.3-pytorch-1.10-deepspeed-0.8.3-gpu-0.21.2
-    cpu: determinedai/environments:cuda-11.3-pytorch-1.10-deepspeed-0.8.3-gpu-0.21.2
 resources:
   slots_per_trial: 2
   # max_slots: 8
