+++ ds_autotuning_prototype/hf_trainer_api/language-modeling/run_clm.py
@@ -41,7 +41,7 @@ from det_callback import (
     get_ds_config_path_from_args,
     replace_ds_config_file_using_overwrites,
 )
-from determined.pytorch.deepspeed import dsat
+from determined.pytorch import dsat
 from torch.utils.tensorboard import SummaryWriter
 from transformers import (
     CONFIG_MAPPING,
