+++ harness/determined/pytorch/deepspeed/__init__.py
@@ -5,7 +5,5 @@ from determined.pytorch.deepspeed._mpu import (
 )
 from determined.pytorch.deepspeed._deepspeed_context import (
     DeepSpeedTrialContext,
-    overwrite_deepspeed_config,
-    get_ds_config_from_hparams,
 )
 from determined.pytorch.deepspeed._deepspeed_trial import DeepSpeedTrial, DeepSpeedTrialController
