+++ harness/determined/pytorch/deepspeed/_deepspeed_trial.py
@@ -17,7 +17,7 @@ from deepspeed.runtime import dataloader as ds_loader
 import determined as det
 from determined import layers, pytorch, util, workload
 from determined.pytorch import deepspeed as det_ds
-from determined.pytorch.deepspeed import dsat
+from determined.pytorch import dsat
 
 
 # In most cases in which a user disables data reproducibility checks and chooses to return
