--- harness/determined/pytorch/deepspeed/dsat/__init__.py
@@ -1,7 +1,0 @@
-from determined.pytorch.deepspeed.dsat._utils import dsat_reporting_context
-from determined.pytorch.deepspeed.dsat._dsat_search_method import (
-    BaseDSATSearchMethod,
-    DSATTrial,
-    DSATTrialTracker,
-    DSATModelProfileInfoTrial,
-)
